+++ src/raspberrypi/rascsi.cpp
@@ -23,10 +23,10 @@
 #include "command_util.h"
 #include "rascsi_version.h"
 #include "rascsi_interface.pb.h"
-#include "rascsi/executor.h"
-#include "rascsi/response.h"
-#include "rascsi/image.h"
-#include "rascsi/service.h"
+#include "rascsi/rascsi_executor.h"
+#include "rascsi/rascsi_response.h"
+#include "rascsi/rascsi_image.h"
+#include "rascsi/rascsi_service.h"
 #include "rasutil.h"
 #include "spdlog/spdlog.h"
 #include "spdlog/sinks/stdout_color_sinks.h"
