+++ src/raspberrypi/rascsi/rascsi_executor.cpp
@@ -14,16 +14,16 @@
 #include "devices/primary_device.h"
 #include "devices/disk.h"
 #include "devices/file_support.h"
-#include "service.h"
-#include "response.h"
-#include "image.h"
+#include "rascsi_service.h"
+#include "rascsi_response.h"
+#include "rascsi_image.h"
 #include "rascsi_exceptions.h"
 #include "localizer.h"
 #include "command_util.h"
 #include "command_context.h"
 #include "rasutil.h"
 #include "spdlog/spdlog.h"
-#include "executor.h"
+#include "rascsi_executor.h"
 #include <sstream>
 
 using namespace spdlog;
