+++ src/raspberrypi/rascsi/rascsi_image.h
