+++ src/raspberrypi/rascsi/rascsi_service.cpp
@@ -13,7 +13,7 @@
 #include "protobuf_serializer.h"
 #include "localizer.h"
 #include "rasutil.h"
-#include "service.h"
+#include "rascsi_service.h"
 #include <netinet/in.h>
 #include <csignal>
